// src/generated/types/DeviceDto.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

import type { DeviceSpecificationDto } from './DeviceSpecificationDto.ts'

export type DeviceDto = {
  /**
   * @type string
   */
  IPAddress: string
  /**
   * @type string
   */
  name: string
  /**
   * @type string
   */
  labId: string
  /**
   * @type string
   */
  assisstantId: string
  /**
   * @type string
   */
  id: string
  /**
   * @type string, date-time
   */
  addedSince: Date
  /**
   * @type string
   */
  status: string
  /**
   * @type array
   */
  specDetails: DeviceSpecificationDto[]
}